adding usfm3 to header on USFM export.
